software_experiments/requirements.txt
absl-py==0.14.1
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
bidict==0.21.2
cachetools==4.2.4
certifi==2021.10.8
chardet==4.0.0
click==8.0.1
colorama==0.4.4
Distance==0.1.3
faiss-gpu==1.7.2
fast-mosestokenizer==0.0.8.1
fastBPE==0.1.0
filelock==3.0.12
fsspec==2021.10.0
future==0.18.2
google-auth==2.3.0
google-auth-oauthlib==0.4.6
grpcio==1.41.0
huggingface-hub==0.0.12
idna==2.10
joblib==1.0.1
jsonargparse==3.13.1
Markdown==3.3.4
multidict==5.2.0
nltk==3.6.2
numpy==1.21.0
oauthlib==3.1.1
packaging==21.0
pandas==1.1.5
Pillow==8.3.1
portalocker==2.3.1
protobuf==3.18.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycld2==0.41
pycld3==0.22
pyDeprecate==0.3.0
pyparsing==2.4.7
-e git+https://github.com/BramVanroy/pyter.git@c78c3d972f0b0de851a7bccd069d258cd86a5816#egg=pyter3
python-dateutil==2.8.2
pytorch-lightning==1.3.5
pytz==2021.3
PyYAML==5.4.1
regex==2021.7.6
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
sacrebleu==2.0.0
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.7.0
sentence-transformers==2.0.0
sentencepiece==0.1.96
six==1.16.0
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
threadpoolctl==2.2.0
tokenizers==0.10.3
torch==1.9.0
torchmetrics==0.5.0
torchvision==0.10.0
tqdm==4.61.2
transformers==4.8.2
typing-extensions==3.10.0.0
unbabel-comet==1.0.0rc6
urllib3==1.26.6
Werkzeug==2.0.2
yarl==1.7.0

